SQL Exercise 10 Output
Exercise 10:

Define a function that retrieves the contract salary value of an employee that also accepts a second parameter. Let this parameter be a character sequence. 
Evaluate if its value is 'min' or 'max' and based on that retrieve either the lowest or the highest salary. 
If the inserted value is any string value different from ‘min’ or ‘max’, let the function return the difference between the highest and the lowest salary of that employee.